api/pyproject.toml
[project]
name = "slidespeaker-api"
version = "0.1.0"
description = "SlideSpeaker - AI-Powered Presentation System"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "fastapi-cors>=0.0.6",
    "aiofiles>=23.2.1",
    "openai>=1.3.5",
    "elevenlabs>=0.2.27",
    "requests>=2.31.0",
    "pypdf2>=3.0.1",
    "python-pptx>=0.6.21",
    "Pillow>=10.0.0",
    "moviepy>=1.0.3",
    "ffmpeg-python>=0.2.0",
    "redis",
    "loguru>=0.7.3",
    "socksio>=1.0.0",
    "reportlab>=4.4.3",
    "imageio>=2.37.0",
    "asyncio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages.find = {exclude = ["output", "uploads", "slidespeaker", "slidespeaker_api.egg-info"]}

[tool.uv]
sources = { pypi = { url = "https://mirrors.aliyun.com/pypi/simple/" } }
